Split Grammar into multiple, specialised subclasses (a la Parser)
