+++ lib/javascript/parser/expression_parser/infix_parser.rb
@@ -1,10 +1,5 @@
 module Javascript
   class Parser::ExpressionParser::InfixParser
-    RIGHT_ASSOCIATIVE_OPERATORS = %w( ** )
-    LEFT_ASSOCIATIVE_OPERATORS  = %w( / * + - , % == === != !== < <= > >= << >> >>> && || & | ^ )
-    UNARY_OPERATORS             = %w( ++ -- )
-    ASSIGNMENT_OPERATORS        = %w( = += -= *= /= **= *= %= <<= >>= >>>= &= |= ^= &&= ||= )
-    
     attr_reader :prefix, :precedence
 
     def initialize(parser:, prefix:, precedence:)
@@ -13,15 +8,15 @@ module Javascript
 
     def parse_infix
       case
-      when tokenizer.consume(RIGHT_ASSOCIATIVE_OPERATORS) then parse_rightward_binary_operation
-      when tokenizer.consume(LEFT_ASSOCIATIVE_OPERATORS)  then parse_leftward_binary_operation
-      when tokenizer.consume(UNARY_OPERATORS)             then parse_unary_operation
-      when tokenizer.consume(ASSIGNMENT_OPERATORS)        then parse_assignment
-      when tokenizer.consume("?.")                        then parse_optional_chaining
-      when tokenizer.consume(".")                         then parse_property_access_with_dot
-      when tokenizer.consume("[")                         then parse_property_access_with_square_brackets
-      when tokenizer.consume("(")                         then parse_function_call
-      when tokenizer.consume("?")                         then parse_ternary
+      when tokenizer.consume("**")                 then parse_rightward_binary_operation
+      when tokenizer.consume(:binary_operator)     then parse_leftward_binary_operation
+      when tokenizer.consume(:unary_operator)      then parse_unary_operation
+      when tokenizer.consume(:assignment_operator) then parse_assignment
+      when tokenizer.consume("?.")                 then parse_optional_chaining
+      when tokenizer.consume(".")                  then parse_property_access_with_dot
+      when tokenizer.consume("[")                  then parse_property_access_with_square_brackets
+      when tokenizer.consume("(")                  then parse_function_call
+      when tokenizer.consume("?")                  then parse_ternary
       end
     end
 
@@ -78,7 +73,7 @@ module Javascript
           raise SyntaxError
         end
       end
-      
+
       def parse_optional_chaining
         case
         when tokenizer.consume("(")
@@ -108,11 +103,11 @@ module Javascript
         [].tap do |arguments|
           tokenizer.until(:closing_bracket) do
             arguments << parse_argument
-            tokenizer.consume(:comma)
+            tokenizer.consume(",")
           end
         end
       end
-      
+
       def parse_argument
         if tokenizer.consume("...")
           Spread.new parser.parse_expression(precedence: 2)
